pyproject.toml
[tool.poetry]
name = "fel"
repository = "https://github.com/zabot/fel"
authors = ["Zach Anderson <zach.inbox@gmail.com>"]
version = "0.3.2"
description = "A tool for submitting and landing stacked diffs to GitHub"
readme = "readme.md"
license = "GPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.7"
git-filter-repo = "^2.29.0"
GitPython = "^3.1.14"
PyGithub = "^1.55"
PyYAML = "^5.4.1"
yaspin = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "^2.12.0"
flake8 = "^3.9.2"
black = "^21.5b2"

[tool.poetry.scripts]
fel = "fel.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
